+++ src/mx/com/desoft/hidrogas/dao/IListaRefaccionesDAO.java
@@ -3,6 +3,6 @@ package mx.com.desoft.hidrogas.dao;
 import mx.com.desoft.hidrogas.hibernate.HibernateDAO;
 import mx.com.desoft.hidrogas.model.ListaRefacciones;
 
-public interface IListaRefaccionesDAO extends HibernateDAO<ListaRefacciones, Long> {
+public interface IListaRefaccionesDAO extends HibernateDAO<ListaRefacciones, Integer> {
 
 }
