+++ src/mx/com/desoft/hidrogas/dao/ListaReaccionesImplDAO.java
@@ -6,6 +6,6 @@ import mx.com.desoft.hidrogas.hibernate.HibernateImplDAO;
 import mx.com.desoft.hidrogas.model.ListaRefacciones;
 
 @Repository
-public class ListaReaccionesImplDAO extends HibernateImplDAO<ListaRefacciones, Long> implements IListaRefaccionesDAO {
+public class ListaReaccionesImplDAO extends HibernateImplDAO<ListaRefacciones, Integer> implements IListaRefaccionesDAO {
 
 }
